backend/requirements.txt
flask>=3.0.0
reportlab>=4.0.0
SQLAlchemy>=2.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.0
email-validator>=2.2.0
python-dotenv>=1.0.1
pytest>=8.2.0
gunicorn>=22.0.0
